requirements.txt
streamlit>=1.11.1
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=0.21.0
lxml>=5.2.0
openpyxl>=3.0.0
tiktoken